secret-store: move db stuff to secret-store as per Anton's request
