+++ parity/db/rocksdb/mod.rs
@@ -36,8 +36,6 @@ use cache::CacheConfig;
 mod blooms;
 mod migration;
 mod helpers;
-#[cfg(feature = "secretstore")]
-mod migration_secretstore;
 
 pub use self::migration::migrate;
 
@@ -59,21 +57,6 @@ impl BlockChainDB for AppDB {
 	fn trace_blooms(&self) -> &blooms_db::Database {
 		&self.trace_blooms
 	}
-}
-
-/// Open a secret store DB using the given secret store data path. The DB path is one level beneath the data path.
-#[cfg(feature = "secretstore")]
-pub fn open_secretstore_db(data_path: &str) -> Result<Arc<dyn KeyValueDB>, String> {
-	use std::path::PathBuf;
-
-	migration_secretstore::upgrade_db(data_path).map_err(|e| e.to_string())?;
-
-	let mut db_path = PathBuf::from(data_path);
-	db_path.push("db");
-	let db_path = db_path.to_str().ok_or_else(|| "Invalid secretstore path".to_string())?;
-
-	let config = DatabaseConfig::with_columns(Some(1));
-	Ok(Arc::new(Database::open(&config, &db_path).map_err(|e| format!("Error opening database: {:?}", e))?))
 }
 
 /// Create a restoration db handler using the config generated by `client_path` and `client_config`.
